+++ CHANGES.md
@@ -2,6 +2,13 @@
 
 This is a history of changes to [Remix](https://remix.run).
 
+## Unreleased
+
+### Improvements
+
+- Added `Cookie` interface and `createCookie` API for handling cookies, including
+  support for signed cookies
+
 ## 0.10.4 - Wed Jan 27 2021
 
 - Fixed Vercel starter deployments